Move Card component to folder
